LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/http-to-https-redirect-and-register.https-expected.txt


PASS register on a secure page after redirect from an non-secure url 
FAIL register on a non-secure page after redirect from an non-secure url assert_equals: expected "FAIL: SecurityError" but got "OK"
